Refactoring tests
Refactoring tests to improve re-use and add tests for multipart/form payload feature.
